Fixing redudant casting from javac warnings.

+++ src/seed/Utility.java
@@ -229,7 +229,7 @@ public class Utility  {
 			System.arraycopy(source, 0, returnByteArray, 0, source.length);
 			for(int i = source.length; i<requiredBytes; i++)
 			{
-				returnByteArray[i] = (byte)paddingByte;
+				returnByteArray[i] = paddingByte;
 			}
 			return returnByteArray;
 		}
